+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-438](https://github.com/itk-dev/hoeringsportal/pull/438)
+  Improved search styling
+* [PR-436](https://github.com/itk-dev/hoeringsportal/pull/436)
+  Improved search
 * [PR-439](https://github.com/itk-dev/hoeringsportal/pull/439)
   Updated tasks and map page fixture.
 * [PR-437](https://github.com/itk-dev/hoeringsportal/pull/437)